.github/scripts/python/requirements.txt
cbor==1.0.0
PyGithub==1.45
